Phone field changed

+++ wincraft_project/winapp/models.py
@@ -4,6 +4,5 @@ from django.db import models
 class Queries(models.Model):
     name = models.CharField(max_length=100)
     email = models.EmailField(blank=True,null=True)
-    phone = models.PositiveIntegerField(blank=True,null=True)
+    phone = models.BigIntegerField(blank=True,null=True)
     query = models.CharField(max_length=1000)
-
